# docs/bookmark-management-openapi.yaml
openapi: 3.1.0
info:
  title: HarborDeck Bookmark Management API
  version: 1.0.0
  description: >-
    AI-oriented API for managing HarborDeck groups, bookmarks, placements,
    ordering, and quick records. The management token can access protected
    scenes but cannot manage scenes, passwords, WebDAV, backups, or system settings.
servers:
  - url: /
security:
  - managementToken: []
...
paths:
...
  /api/management/v1/icons:
    get:
      tags: [Icons]
      operationId: listManagedIcons
      summary: Search supported icon IDs
      parameters:
        - name: q
          in: query
          schema:
            type: string
            maxLength: 100
            default: ''
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
      responses:
        '200':
          description: Icon IDs
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Icon'
        default:
          $ref: '#/components/responses/Error'
  /api/management/v1/scenes/{sceneId}/groups:
    post:
      tags: [Groups]
      operationId: createManagedGroup
      summary: Create a group in an existing scene
      parameters:
        - $ref: '#/components/parameters/SceneId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/DryRun'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreate'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '200':
          $ref: '#/components/responses/DryRun'
        default:
          $ref: '#/components/responses/Error'
  /api/management/v1/scenes/{sceneId}/groups/order:
    put:
      tags: [Groups]
      operationId: reorderManagedGroups
      summary: Replace the complete group order of a scene
      parameters:
        - $ref: '#/components/parameters/SceneId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/DryRun'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [groupIds]
              properties:
                groupIds:
                  type: array
                  items:
                    $ref: '#/components/schemas/Slug'
      responses:
        '200':
          $ref: '#/components/responses/Mutation'
        default:
          $ref: '#/components/responses/Error'
  /api/management/v1/scenes/{sceneId}/groups/{groupId}:
    patch:
      tags: [Groups]
      operationId: renameManagedGroup
      summary: Rename a group without changing its ID
      parameters:
        - $ref: '#/components/parameters/SceneId'
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/DryRun'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name:
                  $ref: '#/components/schemas/Name'
      responses:
        '200':
          $ref: '#/components/responses/Mutation'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Groups]
      operationId: deleteManagedGroup
      summary: Delete a group with an explicit bookmark disposition
      parameters:
        - $ref: '#/components/parameters/SceneId'
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/DryRun'
        - name: bookmarkDisposition
          in: query
          schema:
            type: string
            enum: [reject, remove, move]
            default: reject
        - name: targetGroupId
          in: query
          schema:
            $ref: '#/components/schemas/Slug'
        - name: targetPosition
          in: query
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          $ref: '#/components/responses/Mutation'
        default:
          $ref: '#/components/responses/Error'
  /api/management/v1/bookmarks:
    post:
      tags: [Bookmarks]
      operationId: createManagedBookmark
      summary: Create a bookmark and at least one placement
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/DryRun'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookmarkCreate'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '200':
          $ref: '#/components/responses/DryRun'
        default:
          $ref: '#/components/responses/Error'
  /api/management/v1/bookmarks/{slug}:
    get:
      tags: [Bookmarks]
      operationId: getManagedBookmark
      summary: Return one bookmark and all placements
      parameters:
        - $ref: '#/components/parameters/BookmarkSlug'
      responses:
        '200':
          description: Bookmark details
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [revision, bookmark]
                properties:
                  revision:
                    $ref: '#/components/schemas/Revision'
                  bookmark:
                    $ref: '#/components/schemas/ManagedBookmark'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Bookmarks]
      operationId: updateManagedBookmark
      summary: Partially update a bookmark; null clears optional fields
      parameters:
        - $ref: '#/components/parameters/BookmarkSlug'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/DryRun'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookmarkPatch'
      responses:
        '200':
          $ref: '#/components/responses/Mutation'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Bookmarks]
      operationId: deleteManagedBookmark
      summary: Delete a bookmark definition and all placements
      parameters:
        - $ref: '#/components/parameters/BookmarkSlug'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/DryRun'
      responses:
        '200':
          $ref: '#/components/responses/Mutation'
        default:
          $ref: '#/components/responses/Error'
  /api/management/v1/bookmarks/{slug}/duplicate:
    post:
      tags: [Bookmarks]
      operationId: duplicateManagedBookmark
      summary: Duplicate a bookmark, by default immediately after each source placement
      parameters:
        - $ref: '#/components/parameters/BookmarkSlug'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/DryRun'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookmarkDuplicate'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '200':
          $ref: '#/components/responses/DryRun'
        default:
          $ref: '#/components/responses/Error'
...
  /api/management/v1/scenes/{sceneId}/bookmarks/batch-move:
    post:
      tags: [Placements]
      operationId: batchMoveManagedBookmarks
      summary: Atomically move bookmarks while preserving their prior relative order
      parameters:
        - $ref: '#/components/parameters/SceneId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/DryRun'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchMove'
      responses:
        '200':
          $ref: '#/components/responses/Mutation'
        default:
          $ref: '#/components/responses/Error'
  /api/management/v1/bookmarks/batch-place:
    post:
      tags: [Placements]
      operationId: batchPlaceManagedBookmarks
      summary: Atomically place bookmarks into one or more scenes
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/DryRun'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchPlace'
      responses:
        '200':
          $ref: '#/components/responses/Mutation'
        default:
          $ref: '#/components/responses/Error'
  /api/management/v1/scenes/{sceneId}/bookmarks/batch-remove:
    post:
      tags: [Placements]
      operationId: batchRemoveManagedBookmarks
      summary: Atomically remove bookmarks from one scene
      parameters:
        - $ref: '#/components/parameters/SceneId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/DryRun'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchRemove'
      responses:
        '200':
          $ref: '#/components/responses/Mutation'
        default:
          $ref: '#/components/responses/Error'
  /api/management/v1/scenes/{sceneId}/groups/{groupId}/bookmarks/order:
    put:
      tags: [Placements]
      operationId: reorderManagedBookmarks
      summary: Replace the complete bookmark order of a group
      parameters:
        - $ref: '#/components/parameters/SceneId'
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/DryRun'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [bookmarkIds]
              properties:
                bookmarkIds:
                  type: array
                  items:
                    $ref: '#/components/schemas/Slug'
      responses:
        '200':
          $ref: '#/components/responses/Mutation'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    managementToken:
      type: apiKey
      in: header
      name: X-HarborDeck-Management-Token
...
  parameters:
    SceneId:
      name: sceneId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Slug'
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Quoted or unquoted revision obtained from the latest state read
      schema:
        type: string
        examples:
          - '"sha256:4e9b000000000000000000000000000000000000000000000000000000000000"'
    DryRun:
      name: X-HarborDeck-Dry-Run
      in: header
      required: false
      schema:
        type: string
        enum: ['true']
  responses:
...
    Error:
      description: Request rejected; inspect error.code
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    Slug:
      type: string
      pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
    Name:
      type: string
      minLength: 1
      maxLength: 200
...
    Icon:
      type: object
      additionalProperties: false
      required: [id, label]
      properties:
        id: { type: string }
        label: { type: string }
    GroupCreate:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        id:
          $ref: '#/components/schemas/Slug'
        name:
          $ref: '#/components/schemas/Name'
        position:
          $ref: '#/components/schemas/Position'
...